Move behaviorLookup and features into a config directory
Just for organization
